fix: refatora base da tela do filme
